[Refactor] CommonMenuBar 함수 이동
